ISSUE #3410 - move around helper functions to stop circular dependency
