+++ frontend/src/v5/ui/routes/dashboard/projects/containers/uploadFileForm/uploadFileForm.component.tsx
@@ -36,7 +36,6 @@ import { RevisionsHooksSelectors } from '@/v5/services/selectorsHooks/revisionsS
 import { ContainersHooksSelectors } from '@/v5/services/selectorsHooks/containersSelectors.hooks';
 import { UploadList } from './uploadList';
 import { SidebarForm } from './sidebarForm';
-import { uploadModalLabels } from './uploadFileForm.helpers';
 import { UploadsContainer, DropZone, Modal, UploadsListHeader, Padding, UploadsListScroll } from './uploadFileForm.styles';
 
 type IUploadFileForm = {
@@ -50,6 +49,38 @@ interface AddFilesProps {
 	files: File[];
 	container?: IContainer;
 }
+
+type UploadModalLabelTypes = {
+	isUploading: boolean;
+	fileCount: number;
+};
+
+const uploadModalLabels = ({ isUploading, fileCount }: UploadModalLabelTypes) => (isUploading
+	? {
+		title: formatMessage({
+			id: 'uploads.modal.title.uploading',
+			defaultMessage: '{fileCount, plural, one {Uploading file} other {Uploading files}}',
+		}, { fileCount }),
+		subtitle: formatMessage({
+			id: 'uploads.modal.subtitle.uploading',
+			defaultMessage: '{fileCount, plural, one {Do not close this window until the upload is complete} other {Do not close this window until uploads are complete}}',
+		}, { fileCount }),
+		confirmLabel: formatMessage({ id: 'uploads.modal.buttonText.uploading', defaultMessage: 'Finished' }),
+	}
+	: {
+		title: formatMessage({
+			id: 'uploads.modal.title.preparing',
+			defaultMessage: '{fileCount, plural, =0 {Add files for upload} one {Prepare file for upload} other {Prepare files for upload}}',
+		}, { fileCount }),
+		subtitle: formatMessage({
+			id: 'uploads.modal.title.preparing',
+			defaultMessage: '{fileCount, plural, =0 {Drag and drop or browse your computer} other {Select a file to add Container/Revision details}}',
+		}, { fileCount }),
+		confirmLabel: formatMessage({
+			id: 'uploads.modal.buttonText.preparing',
+			defaultMessage: '{fileCount, plural, one {Upload file} other {Upload files}}',
+		}, { fileCount }),
+	});
 
 export const UploadFileForm = ({
 	presetContainerId,
